Code Explained
There will be two different cases:'-

a. When no. of elements in the array is even.
b. When no. of elements in the array is odd.

This is because when pair elements to be swapped, one element will be left out when the no. of elements in the array is odd.

'num' is the no. of elements in the array.

CASE A: When no. of elements is odd.

Line 15: Initialize i=0 (points to the first element); j = num-1 (points to the last element of array);
             Condition - Check if 'i' is not equal to 'j' (when i==j, both i & j point to the middle element, the middle element stays at its 
                                original position, so i==j is termination condition)
            Increment 'i' & decrement 'j' to point to the next pair of elements.


CASE B: When no. of elements is even.

              All the elements are paired to be swapped in this case.

Line 20: Initialize i=0 (points to the first element); j = num-1 (points to the last element of array);
             Condition - i < (num/2), since total no. of iterations will be half the size of the array because the elements are paired. 
             Increment 'i' & decrement 'j' to point to the next pair of elements.